omxvideodec: integrate resize component on RPI
It enables the use of ximagesink on RPI.
Or any video sink that does not use GL on RPI.
Total CPU load at 20% for a 640x360 video.
Total CPU load at 60% for 1280x720 video.

Mainly because it does colorspace conversion.
It allows to support more downstream elements (or make them
useable because it drastically reduces CPU load)
Ex: Allow to output RGB16 if downstream does not support I420
and if "video_decode" can only output I420.

Only available on RPI (passing --with-omx-target=rpi to
the configure script) because I only tested on RPI and also
because I'm sure if the resize element is available on other
platform. And for now it's instanciated using its name directly:
"OMX.broadcom.resize"

But you does not necessarly need to have GST_EGL as it's the case
to use "OMX.broadcom.egl_render"

First, omxvideodec tries "egl_render" (if egl available) then if it
fails it directly tries "video_decode" as useall but then if it fails
again it tries "resize" with a compatible downstream colorformat.

When trying the resize component it try to do OMX_UseBuffer
(usefull to use downstream XImage create using XShmCreateImage)
and if it fails to actually use those downstream buffers then it
fallbacks to OMX_AllocateBuffer.

Ex: Allow to have zero-copy with omxh264dec ! ximagesink (OMX_UseBuffer)
but omxh264dec ! tee ! ximagesink will do OMX_AllocateBuffer because
tee does not forward the allocation query.

Signed-off-by: Mart Raudsepp <mart@stipit.com>
Updated to a year of changes in git master